refactor: rename HTTP REST to REST (#247)
# Description

Follow-up for unreleased #142.

[The spec](https://a2a-protocol.org/latest/specification/) uses
[`HTTP+JSON/REST`](https://a2a-protocol.org/latest/specification/#11-httpjsonrest-protocol-binding)
wording and other SDKs use just REST in all names (i.e.
[Python](https://github.com/a2aproject/a2a-python/blob/main/src/a2a/server/request_handlers/rest_handler.py)
and
[Golang](https://github.com/a2aproject/a2a-go/blob/release/spec-v1/a2asrv/rest.go)).
