+++ src/server/express/index.ts
@@ -9,5 +9,5 @@ export { jsonRpcHandler } from './json_rpc_handler.js';
 export type { JsonRpcHandlerOptions } from './json_rpc_handler.js';
 export { agentCardHandler } from './agent_card_handler.js';
 export type { AgentCardHandlerOptions, AgentCardProvider } from './agent_card_handler.js';
-export { httpRestHandler } from './http_rest_handler.js';
-export type { HttpRestHandlerOptions } from './http_rest_handler.js';
+export { restHandler } from './rest_handler.js';
+export type { RestHandlerOptions } from './rest_handler.js';
