+++ src/server/express/rest_handler.ts
@@ -13,16 +13,16 @@ import {
   HTTP_STATUS,
   mapErrorToStatus,
   toHTTPError,
-} from '../transports/rest/http_rest_transport_handler.js';
+} from '../transports/rest/rest_transport_handler.js';
 import { ServerCallContext } from '../context.js';
 import { getRequestedExtensions } from '../utils.js';
 import { HTTP_EXTENSION_HEADER } from '../../constants.js';
 import { UserBuilder } from './common.js';
 
 /**
- * Options for configuring the HTTP REST handler.
+ * Options for configuring the HTTP+JSON/REST handler.
  */
-export interface HttpRestHandlerOptions {
+export interface RestHandlerOptions {
   requestHandler: A2ARequestHandler;
   userBuilder: UserBuilder;
 }
@@ -57,11 +57,11 @@ const restErrorHandler: ErrorRequestHandler = (
 type AsyncRouteHandler = (req: Request, res: Response) => Promise<void>;
 
 // ============================================================================
-// HTTP REST Handler - Main Export
+// HTTP+JSON/REST Handler - Main Export
 // ============================================================================
 
 /**
- * Creates Express.js middleware to handle A2A HTTP+REST requests.
+ * Creates Express.js middleware to handle A2A HTTP+JSON/REST requests.
  *
  * This handler implements the A2A REST API specification with snake_case
  * field names, providing endpoints for:
@@ -81,10 +81,10 @@ type AsyncRouteHandler = (req: Request, res: Response) => Promise<void>;
  * ```typescript
  * const app = express();
  * const requestHandler = new DefaultRequestHandler(...);
- * app.use('/api/rest', httpRestHandler({ requestHandler, userBuilder: UserBuilder.noAuthentication }));
+ * app.use('/api/rest', restHandler({ requestHandler, userBuilder: UserBuilder.noAuthentication }));
  * ```
  */
-export function httpRestHandler(options: HttpRestHandlerOptions): RequestHandler {
+export function restHandler(options: RestHandlerOptions): RequestHandler {
   const router = express.Router();
   const restTransportHandler = new RestTransportHandler(options.requestHandler);
