+++ tck/agent/index.ts
@@ -13,7 +13,7 @@ import {
 import {
   jsonRpcHandler,
   agentCardHandler,
-  httpRestHandler,
+  restHandler,
   UserBuilder,
 } from '../../src/server/express/index.js';
 
@@ -211,10 +211,10 @@ async function main() {
     jsonRpcHandler({ requestHandler, userBuilder: UserBuilder.noAuthentication })
   );
 
-  // Register HTTP+REST handler (new feature - additional transport)
+  // Register HTTP+JSON/REST handler (new feature - additional transport)
   expressApp.use(
     '/a2a/rest',
-    httpRestHandler({ requestHandler, userBuilder: UserBuilder.noAuthentication })
+    restHandler({ requestHandler, userBuilder: UserBuilder.noAuthentication })
   );
 
   // 5. Start the server
